Modified. Check whether an app shortcut exists or not

+++ cleaner.sh
@@ -3,24 +3,28 @@ apps_directory="/usr/share/applications/"
 app_shortcut="$1"
 
 if [[ "${app_shortcut}" != "" ]]; then
-    exclusion_pattern='^Comment\['
-    exclusion_pattern+='|^Name\['
-    exclusion_pattern+='|^NoDisplay'
-    exclusion_pattern+='|^$'
+    if [[ -e "${app_shortcut}" ]]; then
+        exclusion_pattern='^Comment\['
+        exclusion_pattern+='|^Name\['
+        exclusion_pattern+='|^NoDisplay'
+        exclusion_pattern+='|^$'
 
-    echo "[+] Simplifying desktop shortcut: ${app_shortcut}"
-    temp_shortcut="shortcut-$(date +%s).temp.txt"
-    current_lines=$(cat "${app_shortcut}" | wc -l)
-    cat "${app_shortcut}" | grep -vE "${exclusion_pattern}" > "${temp_shortcut}"
-    new_lines=$(cat "${temp_shortcut}" | wc -l)
-    echo "    From ${current_lines} to ${new_lines} lines"
+        echo "[+] Simplifying desktop shortcut: ${app_shortcut}"
+        temp_shortcut="shortcut-$(date +%s).temp.txt"
+        current_lines=$(cat "${app_shortcut}" | wc -l)
+        cat "${app_shortcut}" | grep -vE "${exclusion_pattern}" > "${temp_shortcut}"
+        new_lines=$(cat "${temp_shortcut}" | wc -l)
+        echo "    From ${current_lines} to ${new_lines} lines"
 
-    if [[ "${new_lines}" -lt "${current_lines}" ]]; then
-        echo -n "    " && rm -fv "${app_shortcut}"
-        echo -n "    " && mv -v "${temp_shortcut}" "${app_shortcut}"
+        if [[ "${new_lines}" -lt "${current_lines}" ]]; then
+            echo -n "    " && rm -fv "${app_shortcut}"
+            echo -n "    " && mv -v "${temp_shortcut}" "${app_shortcut}"
+        else
+            echo "    App shortcut is already clean"
+            rm -f "${temp_shortcut}"
+        fi
     else
-        echo "    App shortcut is already clean"
-        rm -f "${temp_shortcut}"
+        echo "[x] File does not exists: ${app_shortcut}"
     fi
 
 elif [[ -e "${apps_directory}" ]]; then
